// ui/src/generated/types.d.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: google_oauth2.go

export interface GoogleUser {
  email: string;
  hd: string;
}

//////////
// source: specs.go

export interface ListSpecsRequest {
  Limit: number /* int32 */;
  Offset: number /* int32 */;
  OrderBy: string;
  OrderDir: string;
  Title: string;
  Team: string;
  Type: string;
  Author: string;
}
export interface Spec {
  id: string;
  title: string;
  status: string;
  authors: string[];
  spec_type: string;
  team: string;
  google_doc_id: string;
  google_doc_name: string;
  google_doc_url: string;
  google_doc_created_at: string /* RFC3339 */;
  google_doc_updated_at: string /* RFC3339 */;
  created_at: string /* RFC3339 */;
  updated_at: string /* RFC3339 */;
  synced_at: string /* RFC3339 */;
}
export interface ListSpecsResponse {
  total: number /* int64 */;
  specs: Spec[];
  limit: number /* int32 */;
  offset: number /* int32 */;
}